Tweak map component
