--- ui/src/design-system/map/constants.ts
@@ -1,5 +1,0 @@
-export const ATTRIBUTION =
-  '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-
-export const TILE_LAYER_URL =
-  'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png'
